`alternate-buffer`: Add option whether to restrict to useful buffers
The behaviour where two consecutive presses of SPC TAB always switch back to the
same buffer may be preferable, hence this option.

It might also make sense to combine `spacemacs-layouts-restrict-spc-tab` and
`spacemacs-useful-buffers-restrict-spc-tab` into one option whether to obey the
frame's buffer predicate, but this would not be backwards-compatible, and
clearly not as flexible.